pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unga-analysis"
version = "1.0.0"
description = "Professional framework for analyzing UN General Assembly speeches"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "SMU Data Team", email = "data-team@smu.edu"}
]
keywords = ["unga", "un", "speech", "analysis", "nlp", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Researchers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "streamlit>=1.28.0",
    "openai>=1.43.0",
    "pypdf>=3.15.0",
    "pdfminer.six>=20221105",
    "python-docx>=0.8.11",
    "pydub>=0.25.1",
    "sqlalchemy>=2.0.0",
    "tiktoken>=0.5.0",
    "langdetect>=1.0.9",
    "deep-translator>=1.11.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "duckdb>=0.9.0",
    "sentence-transformers>=2.2.2",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "httpx>=0.28.0",
    "plotly>=5.0.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/smu-data-team/unga-analysis"
Repository = "https://github.com/smu-data-team/unga-analysis"
Issues = "https://github.com/smu-data-team/unga-analysis/issues"

[project.scripts]
unga-analysis = "src.unga_analysis.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
